6.1.15 Выносим общую логику обновления
